pyproject.toml
[project]
name = "weather-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.12.4",
    "httpx>=0.28.1",
    "mcp>=1.15.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
]
